+++ src/climate_data/extract/__init__.py
@@ -1,21 +1,21 @@
-from climate_downscale.extract.cmip6 import (
+from climate_data.extract.cmip6 import (
     extract_cmip6,
     extract_cmip6_task,
 )
-from climate_downscale.extract.elevation import (
+from climate_data.extract.elevation import (
     extract_elevation,
     extract_elevation_task,
 )
-from climate_downscale.extract.era5 import (
+from climate_data.extract.era5 import (
     download_era5_task,
     extract_era5,
     unzip_and_compress_era5_task,
 )
-from climate_downscale.extract.ncei_climate_stations import (
+from climate_data.extract.ncei_climate_stations import (
     extract_ncei_climate_stations,
     extract_ncei_climate_stations_task,
 )
-from climate_downscale.extract.rub_local_climate_zones import (
+from climate_data.extract.rub_local_climate_zones import (
     extract_rub_local_climate_zones,
 )
