+++ src/climate_data/extract/era5.py
@@ -9,8 +9,8 @@ import yaml
 from rra_tools import jobmon
 from rra_tools.shell_tools import touch
 
-from climate_downscale import cli_options as clio
-from climate_downscale.data import DEFAULT_ROOT, ClimateDownscaleData
+from climate_data import cli_options as clio
+from climate_data.data import DEFAULT_ROOT, ClimateDownscaleData
 
 
 def get_download_spec(
