+++ src/climate_data/generate/__init__.py
@@ -1,24 +1,24 @@
-from climate_downscale.generate.derived_daily import (
+from climate_data.generate.derived_daily import (
     generate_derived_daily,
     generate_derived_daily_task,
 )
-from climate_downscale.generate.historical_daily import (
+from climate_data.generate.historical_daily import (
     generate_historical_daily,
     generate_historical_daily_task,
 )
-from climate_downscale.generate.historical_reference import (
+from climate_data.generate.historical_reference import (
     generate_historical_reference,
     generate_historical_reference_task,
 )
-from climate_downscale.generate.scenario_annual import (
+from climate_data.generate.scenario_annual import (
     generate_scenario_annual,
     generate_scenario_annual_task,
 )
-from climate_downscale.generate.scenario_daily import (
+from climate_data.generate.scenario_daily import (
     generate_scenario_daily,
     generate_scenario_daily_task,
 )
-from climate_downscale.generate.scenario_inclusion import (
+from climate_data.generate.scenario_inclusion import (
     generate_scenario_inclusion,
 )
