+++ src/climate_data/generate/derived_daily.py
@@ -4,9 +4,9 @@ import click
 from dask.diagnostics.progress import ProgressBar
 from rra_tools import jobmon
 
-from climate_downscale import cli_options as clio
-from climate_downscale.data import DEFAULT_ROOT, ClimateDownscaleData
-from climate_downscale.generate import utils
+from climate_data import cli_options as clio
+from climate_data.data import DEFAULT_ROOT, ClimateDownscaleData
+from climate_data.generate import utils
 
 TRANSFORM_MAP = {
     "heat_index": utils.Transform(
