+++ src/climate_data/generate/scenario_daily.py
@@ -8,9 +8,9 @@ import pandas as pd
 import xarray as xr
 from rra_tools import jobmon
 
-from climate_downscale import cli_options as clio
-from climate_downscale.data import DEFAULT_ROOT, ClimateDownscaleData
-from climate_downscale.generate import utils
+from climate_data import cli_options as clio
+from climate_data.data import DEFAULT_ROOT, ClimateDownscaleData
+from climate_data.generate import utils
 
 # Map from source variable to a unit conversion function
 CONVERT_MAP = {
